+++ anywise/__init__.py
@@ -8,6 +8,6 @@ from ._itypes import IGuard as IGuard
 from ._registry import MessageRegistry as MessageRegistry
 from .anywise import Anywise as Anywise
 from .guard import BaseGuard as BaseGuard
-from .publisher import concurrent_publish as concurrent_publish
+from .strategies import concurrent_publish as concurrent_publish
 
 # from . import integration as integration